+++ Robust.Shared/Toolshed/TypeParsers/StringTypeParser.cs
@@ -9,12 +9,25 @@ using Robust.Shared.Utility;
 
 namespace Robust.Shared.Toolshed.TypeParsers;
 
-internal sealed class StringTypeParser : TypeParser<string>
+public sealed class StringTypeParser : TypeParser<string>
 {
     // Completion option for hinting that all strings must start with a quote
     private static readonly CompletionOption[] Option = [new("\"", Flags: CompletionOptionFlags.PartialCompletion | CompletionOptionFlags.NoEscape)];
 
     public override bool TryParse(ParserContext ctx, [NotNullWhen(true)] out string? result)
+    {
+        return TryParse(ctx, out result, partial: false);
+    }
+
+    /// <summary>
+    /// Attempt to parse a string.
+    /// </summary>
+    /// <param name="ctx">The parser context.</param>
+    /// <param name="result">The parsed string, without enclosing quotes.</param>
+    /// <param name="partial">If true, this will succeed even without a matching closing quote. This is intended to be
+    /// by string-like parsers (e.g., resource paths) that want to generate completion options for partially complete strings.
+    /// </param>
+    public static bool TryParse(ParserContext ctx, [NotNullWhen(true)] out string? result, bool partial = true)
     {
         ctx.ConsumeWhitespace();
         if (!ctx.EatMatch('"'))
@@ -33,7 +46,7 @@ internal sealed class StringTypeParser : TypeParser<string>
         }
 
         var output = new StringBuilder();
-        while (ctx.GetRune() is {} r)
+        while (ctx.GetRune() is { } r)
         {
             if (r == new Rune('"'))
             {
@@ -43,6 +56,12 @@ internal sealed class StringTypeParser : TypeParser<string>
 
             if (r != new Rune('\\'))
             {
+                // TODO TOOLSHED
+                // Instead of appending each rune individually, its probably better to append whole substrings.
+                // I.e., use something like:
+                // var quoteIndex = ctx.Input.IndexOf('"', ctx.Index);
+                // var escapeIndex = ctx.Input.IndexOf('\\', ctx.Index, quoteIndex - ctx.Index);
+                // output.Append(ctx.Input.AsSpan()[ctx.Index..escapeIndex]);
                 output.Append(r);
                 continue;
             }
@@ -62,10 +81,18 @@ internal sealed class StringTypeParser : TypeParser<string>
             return false;
         }
 
+        if (partial)
+        {
+            // Partial strings don't require closing quotes.
+            result = output.ToString();
+            return true;
+        }
+
         // We ran out of input before encountering the terminating quote.
         // Either someone is trying to execute an incomplete command, or more likely, they forgot the terminating quote.
         if (!ctx.GenerateCompletions)
             ctx.Error = new StringMustEndWithQuote();
+
         result = null;
         return false;
     }
@@ -95,7 +122,7 @@ public record struct StringMustStartWithQuote : IConError
 public sealed class StringMustEndWithQuote : ConError
 {
     public override FormattedMessage DescribeInner()
-        => FormattedMessage.FromUnformatted($"String must end with a quote (\").");
+        => FormattedMessage.FromUnformatted($"Missing closing a quote (\").");
 }
 
 public sealed class UnknownEscapeSequence(Rune c) : ConError